Fixed bug when having more than 9 config files
the maximum of controller configs were increased to 32, but the datatype
was still u16 (which only has 16 bits). Changing it to u32 should
finally fix all issues with the config files

+++ patcher/ControllerPatcherDefs.h
@@ -218,7 +218,7 @@ enum Controller_Patcher_DPAD_Settings
  */
 typedef struct _HIDSlotData{
 	u16 deviceslot;     /**< deviceslot number */
-	u16 hidmask;        /**< Used HID-Mask */
+	u32 hidmask;        /**< Used HID-Mask */
 }HIDSlotData;
 
 /**
